Moved SupabaseClient and Models to Shared Dir
